backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Image processing and ML
opencv-python==4.8.1.78
paddlepaddle==2.5.2
paddleocr==2.7.0.3
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
pillow==10.0.1

# PDF processing
pypdf2==3.0.1
pdf2image==1.16.3

# Data processing
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development and utilities
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# HTTP client
requests==2.31.0

# Logging
structlog==23.2.0

# CORS
fastapi-cors==0.0.6 